Python Notebook
In this project, first I applied stats model analysis to find the relevant columns for the Logistic Regression by finding out p-values for each feature and did back elimination to remove out of relevant features. 
But 'Totchol' feature has an influence records which is affecting the p-value to greater than the alpha(5%).
So, I have included the 'Totchol' for the analysis.
Where removing the influential record will be done in the next part.