KC_HouseSales_Notebook
I have used tensorflow api for this dataset.  The algorithm which I applied to this dataset is the Gradient Descent algorithm inspired from the crash course by google. Though this algorithm works fine, I didn't get the min root mean squared error using this algorithm by Tweaking the hyperparameters like learning rate, batch size, and steps. The R2 score is almost 50% which is not a good performance to this model. Later will come up with a model using the sci-kit libraries for linear regression and back elimination. Have to work on the significant features of the algorithm.